Add addBook module
