+++ index.js
@@ -1,8 +1,8 @@
-import generateUniqueId from './node_modules/generate-unique-id/generateUniqueId.js';
 import getTimeNow from './modules/timeNow.js';
 import toggleSectionVisibility from './modules/toggleSection.js';
 import getStoredBooks from './modules/getStoredBooks.js';
 import showBooks from './modules/showBooks.js';
+import addBook from './modules/addBook.js';
 
 const myBooks = [];
 const addBtn = document.getElementById('add-btn');
@@ -11,37 +11,9 @@ const listBooksLink = document.getElementById('list-books-link');
 const contactLink = document.getElementById('contact-link');
 const currentTime = document.getElementById('current-time');
 
-// const removeBook = (bookIDToRemove) => {
-//   const myBooksUpdated = myBooks.filter((myBook) => myBook.bookID !== bookIDToRemove);
-//   localStorage.setItem('myBooks', JSON.stringify(myBooksUpdated));
-//   myBooks.length = 0;
-//   myBooks.push(...myBooksUpdated);
-//   window.location.reload();
-// };
-
-const addBook = () => {
-  const bookID = generateUniqueId();
-  const bookTitle = document.getElementById('title').value.trim();
-  const bookAuthor = document.getElementById('author').value.trim();
-  const errorSpan = document.getElementById('error-msg');
-  const errorMsg = 'Both Title and Author must be filled out!';
-
-  if (!bookTitle || !bookAuthor) {
-    errorSpan.textContent = errorMsg;
-    return;
-  }
-
-  myBooks.push({ bookID, bookTitle, bookAuthor });
-  localStorage.setItem('myBooks', JSON.stringify(myBooks));
-
-  showBooks(myBooks);
-
-  // Clear the input fields values
-  document.getElementById('title').value = '';
-  document.getElementById('author').value = '';
-};
-
-addBtn.addEventListener('click', addBook);
+addBtn.addEventListener('click', () => {
+  addBook(myBooks);
+});
 
 currentTime.textContent = getTimeNow();
